net/softnic: fix integer overflow
Fixed integer overflow by casting hard_rate to uint64_t type.

Coverity issue: 195020
Fixes: 8316b9816b79 ("net/softnic: support traffic management")

+++ drivers/net/softnic/rte_eth_softnic_tm.c
@@ -47,7 +47,7 @@
 int
 tm_params_check(struct pmd_params *params, uint32_t hard_rate)
 {
-	uint64_t hard_rate_bytes_per_sec = hard_rate * BYTES_IN_MBPS;
+	uint64_t hard_rate_bytes_per_sec = (uint64_t)hard_rate * BYTES_IN_MBPS;
 	uint32_t i;
 
 	/* rate */
